Fix 2.2.9 Bugs for Release. Command Status, Web Cache, & Web Timing (#2169)
* Fix bug getting the status of a command

The nodejs spec for the special cased promisify version of exec mentions that it throws an object with an error, stdout and stderr. It looks like the object it throws is not an error in an object but rather an error object. So calling .error is incorrect.

This fixes the handling of cancellations.

* Dont retry download if it was already downloaded

This also fixes a callback which was added incorrectly during the async refactor as it did not wait on the callback.

* Cache URL Requests Better

The old web request library was a class which each owned their own axios instance.
This used to work when the cache storage was persistent and stored within VS Code. But now that its a local memory cache, I migrated the class to a singleton so the cache actually gets used by everyone.
For why its not persistent storage, there is a bug in VSCode described here which makes doing that not work well: https://github.com/dotnet/vscode-dotnet-runtime/pull/1540

That was a very subtle bug and rare so I dont want to deal with figuring out if it's better or not in a hot fix of releases, but I figured this was simple enough that it would be ok to fit it in. (Maybe I was a bit optimistic.)

Thank you to @baronfel as well for doing some testing around with the axios cache library.

* Fix timer for downloading files

The file download response type is a 'stream' which means the response gets sent back with an open stream to pipe the file.
This means the recorded end time for the webrequest was for the first few chunks of data and the initial stream pipe handover.
Now it waits to record the data until after that is done.

To simplify the logic, it still records every web request but recognizes if its in a stream mode not to record that data and let the end of the stream handler callback report it.

* fix changelog

* fix interfaces to be aligned for mocks

* fix interfaces for mocks

* fix lint

* fix test

* Dont use Cleanup Interval as it Causes Hang Forever

CleanUp Interval spawns a callback which never finishes probably because it was designed for use in a web browser so when you close it, all of the js process is killed.
This causes our tests to never finish or any code relying on it never exits.

I implemented a different method to clean it up -- with a destroy () method (sadly no easy way to add RAII in JS) for tests.
Otherwise the running function object will realize when its been deleted as its been bound with the (this) scope.

You can bind automatically in es6 using an arrow function but that doesn't seem to bind the correct this ->
there is also the that = this convention with .bind(this) but the simplest approach to me is to bind it by making it part of the scope (aka a class member.)

* fix lint

* copilot fixes

* code cleanup

* Always create a Package to test

* Fix Typo


* Dont consider no dotnet install as failure on lnx

---------

+++ vscode-dotnet-runtime-library/src/Acquisition/WinMacGlobalInstaller.ts
@@ -7,12 +7,7 @@ import * as fs from 'fs';
 import * as os from 'os';
 import * as path from 'path';
 
-import * as versionUtils from './VersionUtilities'
-import { FileUtilities } from '../Utils/FileUtilities';
-import { VersionResolver } from './VersionResolver';
-import { WebRequestWorker } from '../Utils/WebRequestWorker';
-import { getInstallFromContext } from '../Utils/InstallIdUtilities';
-import { CommandExecutor } from '../Utils/CommandExecutor';
+
 import
 {
     DotnetAcquisitionAlreadyInstalled,
@@ -29,19 +24,24 @@ import
     OSXOpenNotAvailableError,
     SuppressedAcquisitionError,
 } from '../EventStream/EventStreamEvents';
+import { CommandExecutor } from '../Utils/CommandExecutor';
+import { FileUtilities } from '../Utils/FileUtilities';
+import { getInstallFromContext } from '../Utils/InstallIdUtilities';
+import { WebRequestWorkerSingleton } from '../Utils/WebRequestWorkerSingleton';
+import { VersionResolver } from './VersionResolver';
+import * as versionUtils from './VersionUtilities';
 
-import { IGlobalInstaller } from './IGlobalInstaller';
+import { CommandExecutorResult } from '../Utils/CommandExecutorResult';
 import { ICommandExecutor } from '../Utils/ICommandExecutor';
 import { IFileUtilities } from '../Utils/IFileUtilities';
 import { IUtilityContext } from '../Utils/IUtilityContext';
-import { IAcquisitionWorkerContext } from './IAcquisitionWorkerContext';
-import { DotnetInstall } from './DotnetInstall';
-import { CommandExecutorResult } from '../Utils/CommandExecutorResult';
 import { getOSArch } from '../Utils/TypescriptUtilities';
-import { RegistryReader } from './RegistryReader';
-import { IRegistryReader } from './IRegistryReader';
-import { util } from 'chai';
 import { SYSTEM_INFORMATION_CACHE_DURATION_MS } from './CacheTimeConstants';
+import { DotnetInstall } from './DotnetInstall';
+import { IAcquisitionWorkerContext } from './IAcquisitionWorkerContext';
+import { IGlobalInstaller } from './IGlobalInstaller';
+import { IRegistryReader } from './IRegistryReader';
+import { RegistryReader } from './RegistryReader';
 
 namespace validationPromptConstants
 {
@@ -68,7 +68,7 @@ export class WinMacGlobalInstaller extends IGlobalInstaller
     private completedInstall = false;
     protected versionResolver: VersionResolver;
     public file: IFileUtilities;
-    protected webWorker: WebRequestWorker;
+    protected webWorker: WebRequestWorkerSingleton;
     private invalidIntegrityError = `The integrity of the .NET install file is invalid, or there was no integrity to check and you denied the request to continue with those risks.
 We cannot verify our .NET file host at this time. Please try again later or install the SDK manually.`;
 
@@ -82,7 +82,7 @@ We cannot verify our .NET file host at this time. Please try again later or inst
         this.commandRunner = executor ?? new CommandExecutor(context, utilContext);
         this.versionResolver = new VersionResolver(context);
         this.file = new FileUtilities();
-        this.webWorker = new WebRequestWorker(context, installerUrl);
+        this.webWorker = WebRequestWorkerSingleton.getInstance();
         this.registry = registryReader ?? new RegistryReader(context, utilContext);
     }
 
@@ -236,8 +236,15 @@ This report should be made at https://github.com/dotnet/vscode-dotnet-runtime/is
     private async downloadInstaller(installerUrl: string): Promise<string>
     {
         const ourInstallerDownloadFolder = IGlobalInstaller.getDownloadedInstallFilesFolder(installerUrl);
-        this.file.wipeDirectory(ourInstallerDownloadFolder, this.acquisitionContext.eventStream);
         const installerPath = path.join(ourInstallerDownloadFolder, `${installerUrl.split('/').slice(-1)}`);
+
+        if (fs.existsSync(installerPath) && await this.installerFileHasValidIntegrity(installerPath))
+        {
+            this.acquisitionContext.eventStream.post(new DotnetFileIntegrityCheckEvent(`The installer file ${installerPath} already exists and is valid.`));
+            return installerPath;
+        }
+
+        this.file.wipeDirectory(ourInstallerDownloadFolder, this.acquisitionContext.eventStream);
 
         const installerDir = path.dirname(installerPath);
         if (!fs.existsSync(installerDir))
@@ -245,7 +252,7 @@ This report should be made at https://github.com/dotnet/vscode-dotnet-runtime/is
             fs.mkdirSync(installerDir, { recursive: true });
         }
 
-        await this.webWorker.downloadFile(installerUrl, installerPath);
+        await this.webWorker.downloadFile(installerUrl, installerPath, this.acquisitionContext);
         try
         {
             if (os.platform() === 'win32') // Windows does not have chmod +x ability with nodejs.
@@ -260,7 +267,7 @@ This report should be made at https://github.com/dotnet/vscode-dotnet-runtime/is
             }
             else
             {
-                fs.chmodSync(installerPath, 0o744);
+                await fs.promises.chmod(installerPath, 0o744);
             }
         }
         catch (error: any)
@@ -322,7 +329,7 @@ Please try again, or download the .NET Installer file yourself. You may also rep
             else if (error?.message?.includes('EPERM'))
             {
                 this.acquisitionContext.eventStream.post(new DotnetFileIntegrityFailureEvent(`The file ${installerFile} did not have the correct permissions scope to be assessed.
-Permissions: ${JSON.stringify(await this.commandRunner.execute(CommandExecutor.makeCommand('icacls', [`"${installerFile}"`]), { dotnetInstallToolCacheTtlMs: SYSTEM_INFORMATION_CACHE_DURATION_MS }))}`));
+Permissions: ${JSON.stringify(await this.commandRunner.execute(CommandExecutor.makeCommand('icacls', [`"${installerFile}"`]), { dotnetInstallToolCacheTtlMs: SYSTEM_INFORMATION_CACHE_DURATION_MS }, false))}`));
             }
             return this.userChoosesToContinueWithInvalidHash();
         }
@@ -466,7 +473,7 @@ Please correct your PATH variable or make sure the 'open' utility is installed s
                     // Remove this when https://github.com/typescript-eslint/typescript-eslint/issues/2728 is done
                     // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
                     error.message = `The installer does not have permission to execute. Please try running as an administrator. ${error?.message}.
-Permissions: ${JSON.stringify(await this.commandRunner.execute(CommandExecutor.makeCommand('icacls', [`"${installerPath}"`]), { dotnetInstallToolCacheTtlMs: SYSTEM_INFORMATION_CACHE_DURATION_MS }))}`;
+Permissions: ${JSON.stringify(await this.commandRunner.execute(CommandExecutor.makeCommand('icacls', [`"${installerPath}"`]), { dotnetInstallToolCacheTtlMs: SYSTEM_INFORMATION_CACHE_DURATION_MS }, false))}`;
                 }
                 // Remove this when https://github.com/typescript-eslint/typescript-eslint/issues/2728 is done
                 // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
